Reverse catch-up, i.e., server-demanded stalling
Previously, if two clients were connected to the same server and one of
them was ahead of the server, the only way to rectify that situation was
for the client to get so far ahead that it stalled, as the server could
only catch up with an ahead client if all clients were ahead. That's
unrealistic. This gives the server the alternate option of demanding
that a client stall. This keeps things nicely in line even with >2
players.

+++ network/netplay/netplay_discovery.c
@@ -142,7 +142,7 @@ bool netplay_discovery_driver_ctl(enum rarch_netplay_discovery_ctl_state state, 
          /* And send it off */
          if (sendto(lan_ad_client_fd, (const char *) &ad_packet_buffer,
             2*sizeof(uint32_t), 0, addr->ai_addr, addr->ai_addrlen) <
-            2*sizeof(uint32_t))
+            (ssize_t) (2*sizeof(uint32_t)))
             RARCH_WARN("Failed to send netplay discovery response.\n");
 
          freeaddrinfo_retro(addr);
